Bump to 2.3.x to avoid dependencies with systemPath
On JAXB 2.2.11, the parent uses `systemPath`.
Which in general is already a bad pattern, but in this case is worse
because it tries to load tools.jar from the current JDK.

This dependency got removed in later versions:
https://search.maven.org/artifact/com.sun.xml.bind.mvn/jaxb-parent/2.3.0/pom

I'm choosing 2.3.0, at least for now, to be as close as possible to the
previously used 2.2.11, but with `systemPath` usage removed.